pyproject.toml
[project]
name = "salary-prediction-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "joblib>=1.5.3",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.56.0",
    "supabase>=2.25.1",
    "uvicorn[standard]>=0.44.0",
]
